benchmark: Align perf arguments with runbooks scripts
The `perf` invocation performed by profile-gitaly.sh currently filters
out child processes and kernel stacks. It may be useful to capture a
more complete profile instead, and this aligns better with how `perf` is
invoked by the scripts [1] in the runbooks repository.

Replace --call-graph=fp with -g, which instructs perf to record both
user and kernel stacks.

Remove the --no-inherit flag so stacks of child Git processes are also
captured.

[1] https://gitlab.com/gitlab-com/runbooks/-/blob/92c1c886d05f2a2cdaf59b757f4a2312f07cf092/scripts/gke/perf_flamegraph_for_all_running_processes.sh#L56