housekeeping: Force housekeeping middleware
RewriteHistory RPC has potential to create large amount of loose files.
Since transaction snapshotting scales poorly with the file count, this
is problematic. And although next housekeeping job would take care of
the loose refs, it would still degrade the performance until the next
housekeeping. Therefore, we need to force the housekeeping right after
the RewriteHistory is run.
In this MR we created a list of RPCs that can force the housekeeping,
even if the interval has not yet been reached. This also allows us to
expand the solution later for other RPCs.