middleware: Dry-run transaction snapshotting
Currently we are logging file and directory counts of a repository
when we are taking the transaction snapshot. Since transaction scales
poorly with high file counts, we want to be able to know these counts
prior to enabling transactions.
We are adding a dry run middleware, which walks over the directory and
calculates snapshot stats without creating the actual snapshot, thus
not affecting the performance. With these stats, we will be able to
get an overview of the repository shapes in our production nodes before
enabling transactions.