storagemgr: Optimize recovery middleware
Currently recovery middleware keeps an internal state to see
if a partition is recovered to not attempt the recovery again.
But each service restart would reset this in memory state and
cause starting unnecessary transaction snapshots. Actual log
application would not occur as there would be no log to apply,
but creating the snapshot themselves are introducing unnecessary
overhead. Especially right after a service restart.
We are introducing an optimization to this flow to check if the
given partition has any pending WAL entries before creating a
transaction.