raft: Add E2E test for AddNode operation
Add end-to-end tests for the AddNode operation to verify if the replicas
are started and are becoming part of the cluster. Just to note while the
leader will establish initial communication the destination replica will
be in StateProbe mode due to some membership issues which are yet to
resolved. This MR just focuses on bootstrapping and setting up initial
communication.