storage: Add method to explicitly ask for storage root
Once the relative paths of original and snapshot repositories are
aligned, the storage path of a repository becomes dynamic. Depending
on whether the repository is a snapshot, its storage can be either:

- The root storage (e.g. `gitaly-xxx/yyy/storages.d/default`)
- A snapshot file system root (e.g.
  `gitaly-xxx/yyy/storages.d/default/+gitaly/staging/zz/snapshots/1`)

When working with a snapshot, we sometimes need to explicitly query
the original repository’s storage root. This commit adds a method and
an option to support this. When using the method or passing the
option, we will always return the root storage.