benchmark: Replace Go program with k6
Previously, we introduced a custom Go program which took a workload
definition as input, and orchestrated `ghz` invocations to generate a
concurrent workload mix. Rather than continuing down this path, a better
option from a maintenance and familiarity perspective is to replace the
use of `ghz` with `k6`.

k6 already allows concurrent workloads and has built-in support for gRPC
requests. It's also used elsewhere in GitLab (e.g. for the GPT and CPT
performance tests).

Remove the custom Go program and replace it with k6. We install a custom
k6 binary using the `xk6` utility as we may need to use the xk6-exec
extension.