benchmark: Replace bcc with bpftrace
bcc and bpftrace are both frontends to BPF. We're currently installing
some of the bcc tools from source, however this results in some
nonfunctional tools (like offcputime) which need additional workarounds
to fix. There exists a prebuilt bcc package in the Ubuntu 22.04
repositories called `bpfcc-tools`, but that also includes broken tools.

Apparently this was a known issue with the bcc tooling, so instead of
trying to wrangle a working installation, let's replace it with
bpftrace. Another reason for this decision is that the Gitaly production
fleet is being upgraded from 20.04 to 22.04, and will also face the same
problems with bcc. Using bpftrace would mean we can use consistent
tooling across the benchmark and production machines.

Remove the tasks that compile bcc, and replace it with a prebuilt
bpftrace binary. Remove the bcc tool invocations in profile-gitaly.sh
until we can replace them with bpftrace equivalents.