benchmark: Restart Gitaly before running tests
If a Gitaly process is running but its on-disk binary gets swapped (for
example from executing configure-benchmark-instance), perf will no
longer be able to extract its debug symbols. When this happens, the
resulting flamegraphs will contain an "[unknown]" label for every stack
frame.

Mitigate this by restarting the Gitaly systemd service before executing
the tests. This ensures that the running Gitaly process is always loaded
from the current on-disk binary.