diff: Error in DiffBlobs if whitespace is ignored with blob_pairs
The upstream change b55e6d36eb (diff: ensure consistent diff behavior
with ignore options, 2025-08-08) modified the behaviour of git-diff(1)
such that it returns no raw output lines if the patch body is also
empty.

We worked around this issue for the CommitDiff RPC [1] by chaining
together git-diff-tree(1) and git-diff-pairs(1) to restore the old
behaviour. This was possible because the OIDs provided to that RPC were
treeish, and such could be used as input to git-diff-tree(1).

In DiffBlobs however, the inputs are either pairs of OIDs in the
blob_pairs field, or a set of ChangedPaths structs in the raw_info
field. These inputs are mutually exclusive, and we expect clients to
migrate to use raw_info as a more efficient means of diffing blobs.

The behavioural change in DiffBlobs only occurs when blob_pairs are
provided in conjunction with ignoring whitespace. This is an uncommon
use case that we can try to stop supporting here. With the
aforementioned upstream change, it's unclear whether we'll ever be able
to support this combination going forward.

Return an InvalidArgument error if the client specifies both blob_pairs
and an ignore whitespace option.

[1] https://gitlab.com/gitlab-org/gitaly/-/merge_requests/8093