trace2hooks: Use indexed names for child_start trace2 events
child_start events in trace2 show us commands that processes like
receivepack shells out to. The problem with the way we are currently
logging the field names is that we use the entire command name, which
could include full filepaths.

This poses a potential problem for log indexers as it results in very
high cardinality for log fields.

Instead, change the approach to use indexed fields where we use a
placeholder name for child processes with child_0, child_1 etc. and then
use separate log entries to log the full command name.