benchmark: Define testdata in repositories config
The k6-benchmark.js script currently hardcodes OIDs, references, and
paths for use in gRPC requests. The values are only applicable to a
single repository (the `git` source) and are static across the lifetime
of the test.

In order to support multiple repositories, we should remove the
hardcoded test inputs and allow them to be defined alongside the
repositories in the config.yml file. The RPCs we're currently calling
require four types of inputs:

- commitish OIDs (i.e. commits)
- references
- filenames
- directories

Modify the config.yml.example to include a few examples of each type.
Modify the Ansible script to convert the `repositories` block into JSON
and write it to the client machine. A subsequent commit will modify the
k6-benchmark.js script to consume the config file instead of hardcoding
the values.