migrator: Reduce the backoff limit to 6hrs
The reftable migrator is an opportunistic migrator which attempts to
migrate repositories whenever there are incoming read requests. The
migrator cancels any ongoing migrations when a corresponding write
request is received.

The max delay for this has been set to 32hrs, which is very conservative.
Especially since one Gitaly node can only process one migration request
at a time. So let's reduce this to 6hrs.

Add a missing log field while we're here.