benchmark: Select test data from repositories.json
In a previous commit, we extended the `config.yml` to allow test data to
be defined for each repository. The `repositories` block of the config
was provided to the client machine as a `repositories.json` file.

We can now remove the hardcoded RPC inputs from the benchmarking script,
and replace them with values that are randomly selected from the JSON
file.