diff: Modify parsePaths to take a callback
parsePaths is a helper function that can parse the raw output format
produced by git-diff-tree(1). It's currently used in the
FindChangedPaths RPC to return structured information to the caller, and
takes an RPC-specific diffChunker as input.

Generalise this function by replacing the diffChunker with a callback
function instead. A subsequent commit will use this helper to parse the
output of git-diff-tree(1) in a different context.