diff: Send empty patch bodies for whitespace-only changes
The upstream change b55e6d36eb (diff: ensure consistent diff behavior
with ignore options, 2025-08-08) modified the behaviour of git-diff(1)
such that it returns no raw output lines if the patch body is also
empty.

This triggers an edge case where if this RPC is called with whitespace
ignored, and the entire diff is only comprised of whitespace, we will
stop returning an entry for it. Previously, we would return an entry for
a diff containing only whitespace _even if it's ignored_. This enabled
the client (namely Rails) to still render the file in the UI with an
indicator that only whitespace was changed.

To retain this behaviour, we execute git-diff-tree(1) without ever
ignoring whitespace, which gives us a complete "diff manifest". As we
iterate over the patches produced by git-diff(1) in eachDiff(), we prune
entries from the manifest which match the diff currently being
processed. What remains at the end is a manifest that contains only
patches consisting of whitespace changes, which we iterate over and send
to the client with an empty patch body.