housekeeping: Use original repo relative path as middleware key
We use a repository’s relative path as the key in the `repoActivity`
map. If the path is present and marked as active, we skip running
housekeeping. However, snapshot repositories do not share the same
relative path as their originals. For example:

- Original repo: `@hashed/58/8c/588c2b0xxx.git`
- Snapshot repo:
  `+gitaly/staging/xxx/snapshots/y/@hashed/58/8c/588c2b0xxx.git`

If we use the snapshot’s relative path as the key, housekeeping may
be scheduled repeatedly instead of being skipped.

There is an ongoing effort to align these paths in
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/8013.

This commit ensures we always use the original repo’s relative path
as the key.