server: Use housekeeping middleware only for external services
When transactions are enabled, it overrides the relative path of
the repository to point to the snapshot directory. Which would make
the housekeeping related caches unusable. We should use the original
relative path when transaction is enabled, but when request is routed
through hook back to gitaly, the original repository is not in the
context anymore. Therefore, housekeeping should not be configured in
the internal gRPC server used for hooks.