Revert "housekeeping: Use original repo relative path as middleware key"
The housekeeping middleware is above the transcation middleware, and thus
the transaction isn't present there to extract the original repository.
The original issue we needed to fix was that the housekeeping middleware
was configured in the internal gRPC server used by hooks. Since that is
fixed now, let's revert this one as this isn't needed.

This reverts commit 01badb46b220a7356eede538f5caa49120aff10d.