serve: Start assignment worker as async operation
Previously partition assignment worker was configured to run only when
raft is enabled. But because partition assignment is now prerequisite
for partition backups as well, we need to run this migration even if
raft is disabled. At the same time this has potential to delay the
startup times depending on the amount of partitions. Therefore, this
commit also changes this process to run async.