service/diff: Validate the DiffBlobs `rawInfo` request field
The DiffBlobs RPC `rawInfo` request field specifies the filepair
information that gets fed to the underlying git-diff-pairs(1) in the raw
format. Missing information in this request field may result in the Git
process crashing.

Validate the `rawInfo` field to ensure the incoming request matches what
git-diff-pairs(1) expects. If validation fails, an invalid argument
error is returned instead of an internal error that resulted from Git
crashing.