migration: Warn when leftover migration garbage folder exists
The leftover migration garbage folder can take up disk space.
Although it does not grow over time, it is useful to notify
administrators of its existence and encourage cleanup.

This commit adds report logic that prints a warning log for each
storage if the garbage folder exists when Gitaly starts.