grpc: Call CloseAndRecv when EOF received
Send sends a request message to the server.  The client may call Send
multiple times to send multiple messages to the server.  On error, Send
aborts the stream.  If the error was generated by the client, the status
is returned directly.  Otherwise, io.EOF is returned, and the status of
he stream may be discovered using CloseAndRecv().

There are two places where we use ClientStreamingClient that can
benefit from this error handling.
- setRepoHooks actually already handles this by ignoring EOF which
then calls CloseAndRecv, but here handling is bit implicit, so I have
refactored it to be more explicit error handling.
- sendToNode in the raftmgr can also benefit from EOF handling.

There are other RPCs where we use bidi streams such as UploadArchive
etc. But those RPCs are running Send and Recv concurrently, and calling
Recv within Send goroutine would not be safe as stated in grpc doc:

It is safe to have a goroutine calling SendMsg and another goroutine
calling RecvMsg on the same stream at the same time, but it is not
safe to call RecvMsg on the same stream in different goroutines.