benchmark: Allow different Gitaly revisions
Previously, the user could specify a single Gitaly revision to execute
benchmarks against. The same revision would be cloned on each Gitaly
node as well as the client machine (to consume the protobuf
definitions).

To perform benchmarks between different versions of Gitaly, it would be
convenient if a different revision could be specified for each target
instance.