test: Check specific Git version
ebb834bcf (Fix ObjectInfo parsing while handling submodule entries,
2025-09-08) introduced a fix to stop asserting the ":sub" suffix of the
revision when dealing with submodules, due to a change in
Git >= 2.51.0.

In addition to checking if the GitMaster feature flag is enabled, we
should check the specific version of Git currently in use. We can be
executing this test under newer versions of Git despite the flag being
disabled, for example during a nightly test.