benchmark: Allow reftables to be tested
Adds a new `reference_backend` configuration option for each repository
which allows for the "files" or "reftable" backend to be specified. The
repository will be cloned with the --ref-format argument using the
desired backend.

Since the prepare-repos-vm runs an older version of Ubuntu, we need to
manually compile a more recent version of Git to use the --ref-format
argument.