benchmark: Build Git with frame pointers preserved
We discovered [1] that the Git binary being used by Gitaly in the
benchmarking instances are not being compiled with frame pointers. This
affects how our profiling results are presente and differs from how Git
is built via Omnibus.

Git is indirectly built via Gitaly's Makefile. We expose a parameter
called GIT_APPEND_BUILD_OPTIONS which allows Git build options to be
threaded through to the Makefile of each Git version we build. In the
Omnibus build [2], we explicitly set GIT_APPEND_BUILD_OPTIONS to include
the -fno-omit-frame-pointer compiler flag. For some reason, we don't
set this directly in Gitaly's Makefile but rely on consumers of the
Makefile to set it.

We can do the same for benchmarking by modifying the "Build Gitaly"
Ansible task. Here, we also include the -g flag so debug symbols are
included, in case we ever need them.

Testing with a similar methodology as [1] on a Gitaly instance
provisioned via the Ansible task demonstrates that the Git binary is now
being compiled with both debug symbols and frame pointers:

$ hostname
jliu-2025-09-24-gitaly-transactions-enabled
$ sudo systemctl restart gitaly
$ ls /var/opt/gitaly/run
gitaly-188513
$ sudo readelf --debug-dump=info /var/opt/gitaly/run/gitaly-188513/gitaly-git-master | grep -o -e 'DW_AT_producer.*' | sort | uniq -c | sort -rn | head
    405 DW_AT_producer    : (indirect string, offset: 0x1aa5): GNU C99 11.4.0 -mtune=generic -march=x86-64 -g -O2 -std=gnu99 -fno-common -fno-omit-frame-pointer -fasynchronous-unwind-tables -fstack-protector-strong -fstack-clash-protection -fcf-protection

[1] https://gitlab.com/gitlab-org/gitaly/-/issues/6732#note_2776156999
[2] https://gitlab.com/gitlab-org/omnibus-gitlab/-/blob/9a7d25113690b9c1173edf150dec9de6ee76c734/config/software/gitaly.rb#L77