makefile: Remove Go build tags for gitaly auxiliary binaries
It was mentionned during an investigation into transaction
performance that the AWS SDK Go module was included into
the build of some Gitaly auxiliary binaries, such as
gitaly-hooks and gitaly-lfs-smudge.

This Go module has `init` functions that initialize global
variables that are megabytes in size, and initializing them
takes quite some time also. Since `gitaly-hooks` is a binary
that is spawned very often by Gitaly, it means those variables
are initialized for almost every transction. It has been
identified as a source of performance degradation for transaction.

As it turns out, this module is included in the build as a
transitive dependency of LabKit, which is a library that
provides helper functions to initialize a tracer for distributed
tracing. In production, we use Stackdriver (Google Monitoring) as
the tracing backend. The library used by LabKit to initialize
a tracer exporter for Stackdriver has an indirect dependency
on the AWS SDK.

This commit removes the build tags for all Gitaly auxiliary
binaries. Doing so, it disables distributed tracing for those
binaries. But it fixes this performance degradation.

Interestingly, this fix has already been merged a couple years
ago.

See here: https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4451

But over time changes were made to the Makefile which inadvertently
re-introduced the issue.

As for the drawback of disabling distributed tracing, an MR already
opened propose to migrate distributed tracing from OpenTracing to
OpenTelemetry. Doing so, it uses the new OpenTelemetry Stackdriver
exporter, which has no dependency on AWS SDK. Hopefully, once this
MR is merged, we could enable back distributed tracing without
performance degradation.

References:
* https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4451
* https://gitlab.com/gitlab-org/gitaly/-/merge_requests/8112

* ...