test: Remove invalid symref deletion test
As discussed in [1], the transaction manager test case that exercises
the deletion of an invalid symref (one that points to a non-existent
reference), is not a situation that we would experience in practice.
Gitaly doesn't create its own symrefs, so it would be unlikely that it
would be asked to delete an invalid symref.

As the upstream change ebb834bc (Fix ObjectInfo parsing while handling
submodule entries, 2025-09-08) has introduced a breaking change with how
git-update-ref handles dangling symref deletions, this test case will
fail and block the release pipeline.

Delete the test. If a failure were to occur, the stderr output of the
git-update-ref command is already relayed to the client, so we would be
able to see this in the logs.

[1] https://gitlab.com/gitlab-org/gitaly/-/merge_requests/8161#note_2767808133