Remove prepare-vm
Removed the complexity of starting a prepare-vm to create snapshots
and attach disks to the gitalies. Each gitaly will just directly clone
in parallel the repositories they need to disk.

Remove the use of the startup script from gitaly
in favour of offloading to ansible for better logs.

Remove setup repo script.