storagemgr: Refactor mock setup
We will access partition and KV data through storage manager
in the following commit, but that would make the tests fail
because stub PartitionFactory doesn't have KV implemented for
mockTransaction.
This commit extends the existing mock structure to add KV
capabilities as well as changing mockNode to be type of
*nodeimpl.Manager to be able to call Close() on it, otherwise
we would have goroutine leak from the newStubPartitionFactory.