benchmarks: Add SETUP parameter to use Kubernetes
Gitaly benchnmarks allow to run a set of benchmarks against
a Gitaly server and to monitor those benchmarks to extract
results that can later be analyzed.

Currently, benchmarks are always run against a Gitaly instance
running on VMs. Whith Gitaly on Kubernetes soon to be officialy
supported, there is a need to benchmark Gitaly running on Kubernetes.

This commit adds the necessary logic to createa Kubernetes cluster
on which Gitaly could run. A new parameter called `SETUP` is
introduced, which can take two values: `vm` or `kubernetes'.

When `kubernetes` is used, a Kubernetes cluster is created
instead of standalone virtual machines.

When `vm` is used, the current setup is deployed as is.

Upcoming commits and MRs will introduces changes to configure
the cluster to run Gitaly in a Pod. This commit only adds the
necessary changes to create the cluster.

References:
* https://gitlab.com/gitlab-org/gitaly/-/issues/6968