Bundle-URI: Remove storage name from bundle path
When storing a bundle to storage, Gitaly generates a
bundle name using the tuple (storage_name, relative_path).

This scheme does not work well when using Gitaly Cluster
because the storage name is unique per node, which results
in the same bundle being genearted `n` number of times, where
`n` equals the number of nodes in the cluster. This situation
wastes storage space.

This commit adds a new entry in the gRPC metadata of a request
in Praefect, which contains the virtual storage name. Now, when
generating a bundle, we look into the metadata to see if this
entry is set. If so, we use the virtual storage name in the bundle
path. If not, we use the repository storage name (same as current).

Now, each node of the cluster can share the same bundle URI
for the same repository.

Relevant resources:

* https://gitlab.com/gitlab-org/gitaly/-/issues/6603