auth: Add helper to determine if a request is authenticated
Add a helper that checks whether or not the "username" metaata field in
the incoming context is set. If it is, we assume that this is an
authenticated request from the Rails client.

This helper should not be treated as bulletproof from a security
perspective, since the metadata field could be spoofed.