snapshot: Log active snapshots by relative path
When debugging transactions performance, it can be useful to know how
many active snapshots exist for a given repository. We currently only
track active _shared_ snapshots, not exclusive ones created in response
to a mutator RPC.

Add a new map guarded by the snapshot.Manager.mutex to track _all_
active snapshots for a given key. The key is the first relative path
provided to snapshot.Manager.GetSnapshot(), which represents the target
repository. This can represent how many shared readers and writers
exist for a directory tree on disk and the degree of contention on
filesystem shared structures.