raft: Remove term and index from JoinCluster Request
The term and index values would conflict when the node boots up, as both
are set to 1. But during the join process, we pass in the leader’s term
and index. To avoid this mismatch, we should exclude the term and index
during the join, as they will be updated anyway once the node becomes
part of the cluster.