ci: Auto-run rails specs for Git version updates
Previously, rails specs had to be triggered manually for the  merge
requests. This commit adds a new rule to automatically run rails specs
when the 'run-rails-specs' label is present on the MR. This label will
be added by the renovate bot whenever it creates a MR to update the Git
version. This would ensure that any regressions introduced by a new Git
version are caught before merging.