benchmarks: Add GCP cleanup functions in Golang
When running benchmarks, resources are created in GCP, such as
instances and disks.

Those resources are created using Terraform, and a script exists
to manually destroy the Terraform environment. However, since
destroying the environment is a manual step, it's possible to
forget the step, and this can cause cloud costs to rise unnecessarily.

This commit introduces a GCP Run function in Golang. This function
will run at an interval and delete resources that are older than a
certain threshold.

This change addresses the need by:

* https://gitlab.com/gitlab-org/gitaly/-/issues/6942