config: Add unauthenticated concurrency limit
It's useful to treat unauthenticated requests different from
authenticated ones. Let's add a section in the concurrency limiter
config that allows this.

To do this, we can create a new ConcurrencyLimits struct and embed it
within the existing Concurrency struct so that by default the limits are
defined for authenticated requests, with an added field to define the
limits for unauthenticated requests.

This code also restructures the validation code.