benchmark: Add area plot for RPC latency
When transactions are enabled, we measure both the total latency of an
RPC request as well as the latency for creating a snapshot for the
transaction to operate against. In the future, we may measure even more
aspects of transactions overhead, such as conflict-checking and WAL
flushing.

Add an area plot to the benchmarking framework which clearly shows the
overhead of snapshot creation on the overall RPC latency. As we measure
more aspects of a request, the graph can be further broken down.