git/stats: Expose optional multi-pack index chunks
The multi-pack index is structured around so-called "chunks". Each chunk
has a specific type that tells the reader what kind of data the chunk
contains. This format allows packfiles to add new chunk types over time
in a backwards-compatible way, as clients can simply ignore chunks of an
unknown type.

Git has added a couple of new chunk types over time. In contrast to the
"base" chunks, those are entirely optional and may only be generated
when certain flags are passed to git-multi-pack-index(1). For us it can
be of importance though whether or not those chunks exist, as the
existence of optional chunks may explain observed behaviour.

Luckily, determining the absence of chunks is rather easy to do: the
multi-pack index header encodes the format version and the number of
chunk types. The header is then followed by a table of contents that
is made up of pairs of chunk "names" and their respective offsets. So
all we have to do is to extract those chunk names to figure out whether
a given chunk type exists.

Extract this information and store it in the repository info structure
so that we can expose the data via the `RepositoryInfo()` RPC in the
next commit.

Note that we only expose the existence of optional chunk types. There
isn't really much of a reason to expose the existence of mandatory
chunks, as Git would fail reading these multi-pack indices anyway.