benchmark: Capture Go CPU profiles
The benchmarking experiments are missing goroutine-level tracing and
profiling which can be useful in bottleneck discovery. Since the
Prometheus collector is already enabled on the Gitaly instance, we can
simply retrieve the profiles from the `/debug/pprof` endpoint.

Modify the profile script to capture 60 seconds of CPU traces and
profile data. The profile data must be captured separately, but will be
automatically embedded into the "trace.out" file. A tool like gotraceui
or `go tool trace` can consume the profile and trace directly from the
one file.