grpc: Return RESOURCE_EXHAUSTED when rate-limited by Rails
Gitaly may invoke Rails endpoints such as POST /internal/allowed during
the course of RPC execution. Hook-based RPCs for example will query the
endpoint to perform an authentication check.

When clients inadvertently send too many concurrent requests to a GitLab
API that invokes one of these Gitaly RPCs, Rails may return an HTTP
429 error when Gitaly tries to perform the authentication check. When
this happens, the error is treated as an internal failure, and Gitaly
returns an INTERNAL status code in the gRPC response.

INTERNAL errors directly contribute to the error rate calculations for
the Gitaly service and have a significant impact on its apdex score.
More importantly, the INTERNAL status code is not representative of the
actual error and doesn't give the caller the opportunity to "back off".

Let's check explicitly for 429 errors and propagate this information up
the call chain. Do this by embedding client.RailsRateLimitedError within
hook.NotAllowedError so callers can invoke errors.Is to handle them
differently. The handling is performed in the statushandler middleware
which returns a RESOURCE_EXHAUSTED status if the error is present in the
chain.

A small change was required in the UserDeleteBranch handler to prevent
it from prematurely returning an INTERNAL error.