benchmarking: Fix permissions when adding refs
The "Add additional refs if required" Ansible task executes
git-update-ref --stdin to add refs in bulk. The paths in the shell
script are relative to the /mnt/git-repositories directory, and the
cloned repositories are owned by the `git` user and group.

The task currently fails due to two issues:

1. The task executes under the gitaly_bench user as default, which
   causes Git to complain about dubious ownership of the repository.
2. It's possible we're not in the /mnt/git-repositories directory, so
   the relative paths won't make sense.

Modify the task so we change to the correct directory and assume the
`git` user.