migration: Bump leftover migration ID to rerun
The leftover migration previously used ID 1, and it has already run
across the gitlab-org group. During some CNY incidents, Gitaly
transactions were paused and the leftover migration was disabled.

Now that Gitaly transactions on CNY have been re-enabled and we want to
turn leftover migration back on, the existing migration will not run
again because it has already been marked as completed.

To ensure that the migration runs a second time, we need to bump the
migration ID so that Gitaly recognizes it as a new migration task.

This commit increments the leftover migration ID to trigger it to run
again.