Add `pagination_cursor` to `ListCommitsResponse`
The `ListCommits` RPC's `ListCommitsRequest` accepts `pagination_params`
containing a `page_token` and `limit` but the `ListCommitsResponse` does
not return the `next_cursor` to be used as the next `page_token`.

- https://gitlab-org.gitlab.io/gitaly/#gitaly.ListCommitsRequest
- https://gitlab-org.gitlab.io/gitaly/#gitaly.PaginationParameter
- https://gitlab-org.gitlab.io/gitaly/#gitaly.ListCommitsResponse
- https://gitlab-org.gitlab.io/gitaly/#gitaly.PaginationCursor

This change adds a `pagination_cursor` field to `ListCommitsResponse` to
return the SHA of the last commit in the response. This cursor can be
used to fetch the next page in a follow-up request.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/555079+

Changelog: added