benchmark: Delete results after run
The benchmarking tool currently does not delete directories created
under the /tmp directory for storing run-specific outputs. Since the
entirety of these results are tarred and downloaded to the local
machine, there's no need to keep these around the client and Gitaly
instances.

Add Ansible tasks to remove the run directories from the Gitaly
instances, and the result directory and tarball from the client.