proto: Selectively initialise the protoregistry
Gitaly maintains a protoregistry to facilitate runtime introspection of
its protobufs. We typically use it as part of middleware, for example in
transactionalizeRequest() to check if an incoming request should be run
with transactions.

The protoregistry is implicitly initialised via an init() function in
the protoregistry package which sets the global GitalyProtoPreregistered
variable. Since we build multiple binaries which make and receive gRPC
calls, and our tests make use of test servers, this is a convenient
catch-all.

During the recent Gitaly benchmarking efforts, we noticed that
significant time was spent inside the `reference-trans` binary to
initialise the protoregistry. `reference-trans` is a symlink of the
`gitaly-hooks` binary which Git executes in the course of processing
reference modifications in the repository. `gitaly-hooks` then requests
various Gitaly RPCs under the HookService [1] to complete the change.

The write operations executed in the benchmarks were typically cheap
operations like WriteRef and DeleteRefs. The total duration of the Git
invocation was minimal, which prevents us from amortising the cost of
initialising the entire protoregistry. Local testing shows that loading
all protobufs takes ~1.8ms, while loading just hook.proto takes ~300us.

Add a conditional to the init() function to check if we're running as
the `gitaly-hooks` binary, and only load the appropriate hook.proto
definition if so.

[1] See gitaly-hooks/hooks.go